Added user input tool file
also fixed python tool (whoops)
