+++ oi.py
@@ -9,7 +9,7 @@ from cli_utils import Text
 from tools.tools import function_tool, ToolRegistry
 from conversation import Message, Conversation
 import tools
-from tools.terminal import user_input
+from tools.user_input import user_input
 import locale
 
 class Interpreter:
